pyproject.toml
[tool.poetry]
name = "dmg-cracker"
version = "0.1.0"
description = ""
authors = ["James Davis <jamesecd@gmail.com>"]
readme = "README.md"
packages = [{include = "dmg_cracker"}]

[tool.poetry.dependencies]
python = "^3.11"
pycrypto = "^2.6.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
